seq2seq/requirements.txt
anyio==4.2.0
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
fonttools==4.47.2
fqdn==1.5.1
fsspec==2023.4.0
gitdb==4.0.11
gitpython==3.1.41
idna==3.4
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.7.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter-server-terminals==0.5.2
jupyter-server==2.12.5
jupyter==1.0.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.9
jupyterlab==4.0.11
kiwisolver==1.4.5
lazy-loader==0.3
markupsafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.7.4
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.0
nltk==3.8.1
notebook-shim==0.2.3
notebook==7.0.7
numpy==1.24.4
opencv-python==4.9.0.80
overrides==7.6.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
pipreqs==0.5.0
pkgutil-resolve-name==1.3.10
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywavelets==1.4.1
pywin32==306
pywinpty==2.0.12
pyyaml==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
qtpy==2.4.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
send2trash==1.8.2
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==58.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing-extensions==4.4.0
tzdata==2023.4
uri-template==1.3.0
urllib3==1.26.13
wandb==0.16.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
yarg==0.1.9
zipp==3.17.0